Sync image-picker selection with the uploads list
Deleting the last "previously used photo" left the deleted image stuck in the
picker: the delete handler cleared the selection, the auto-select effect
immediately re-picked the same photo (the uploads cache lagged a tick), and
nothing reconciled once the cache emptied.

Replace the bare auto-select effects on both try-on pages with a shared
useSelectedUploadSync hook that reconciles the preview against the uploads
list (auto-select most recent, advance past a deleted photo, fall back to the
empty state when none remain). It skips reconciliation while generating, since
the selection is then the just-uploaded image not yet in the uploads list.
