+++ app/src/hooks/tryOn/index.ts
@@ -2,3 +2,4 @@ export { useTryOnGeneration } from './useTryOnGeneration'
 export { useTryOnAnalytics } from './useTryOnAnalytics'
 export { useTryOnImage } from './useTryOnImage'
 export { useTryOnWithOtherPhoto } from './useTryOnWithOtherPhoto'
+export { useSelectedUploadSync } from './useSelectedUploadSync'
